feat(project): structure
initial project skeleton
